requirements.txt
appdirs==1.4.3
appnope==0.1.0
attrs==17.4.0
bidict
colorama==0.3.9
coverage==4.4.2
decorator==4.0.4
flake8==3.5.0
hypothesis==3.44.23
ipython==4.0.1
ipython-genutils==0.1.0
mccabe==0.6.1
mock==1.3.0
mypy==0.560
nose==1.3.7
packaging==16.8
path.py==8.1.2
pbr==1.8.1
pexpect==4.0.1
pickleshare==0.5
pluggy==0.6.0
psutil==5.4.3
ptyprocess==0.5
py==1.5.2
pybuilder==0.11.12
pycodestyle==2.3.1
pyflakes==1.6.0
pyparsing==2.2.0
pyreadline==2.1
pyserial==2.7
pytest==3.4.0
pytest-cov==2.5.1
PyYAML==3.12
simplegeneric==0.8.1
singleton
six==1.11.0
tblib==1.3.2
tox==2.9.1
traitlets==4.0.0
typed-ast==1.1.0
virtualenv==15.1.0
